pyproject.toml
[project]
name = "llm-ts-mini"
version = "1.0.0"
description = "Comprehensive system monitoring and anomaly detection dashboard"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "plotly>=5.15.0",
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    "chromadb>=0.4.0",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "seaborn>=0.12.0",
    "matplotlib>=3.7.0",
    "prophet>=1.1.0",
    "pyarrow>=12.0.0"
]

[project.optional-dependencies]
dev = ["pytest", "black", "flake8"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"